moithudeuon

+++ SWP391/src/java/DHTV/ControllerAdmin/DeleteCategoryServlet.java
@@ -6,6 +6,7 @@
 package DHTV.ControllerAdmin;
 
 import DHTV.category.CategoryDAO;
+import DVHT.userdetails.UserDetailsDTO;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.sql.SQLException;
@@ -16,6 +17,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 /**
  *
@@ -37,14 +39,24 @@ public class DeleteCategoryServlet extends HttpServlet {
             throws ServletException, IOException {
         response.setContentType("text/html;charset=UTF-8");
         String url = "";
+        HttpSession session = request.getSession(false);
         try {
-            int CategoryID = Integer.parseInt(request.getParameter("txtCategoryID"));
-            boolean Status = false;
+            if (session != null) {
+                UserDetailsDTO dto1 = (UserDetailsDTO) session.getAttribute("USER");
 
-            CategoryDAO dao = new CategoryDAO();
-            dao.deleteCategoryAdmin(CategoryID, Status);
-            url = "showCategoryServlet";
+                if (dto1 != null) {
+                    if (dto1.getRoleID() == 1 || dto1.getRoleID() == 2) {
+                        int CategoryID = Integer.parseInt(request.getParameter("txtCategoryID"));
+                        boolean Status = false;
 
+                        CategoryDAO dao = new CategoryDAO();
+                        dao.deleteCategoryAdmin(CategoryID, Status);
+                        url = "showCategoryServlet";
+                    }
+                }
+         }else{
+            url = "erorr.jsp";
+            }
         } catch (NamingException ex) {
             log("ShowItemsServlet _ Naming _ " + ex.getMessage());
         } catch (SQLException ex) {
