+++ SWP391/src/java/DHTV/ControllerAdmin/DeletePaymentServlet.java
@@ -6,6 +6,7 @@
 package DHTV.ControllerAdmin;
 
 import DHTV.payment.PaymentMethodDAO;
+import DVHT.userdetails.UserDetailsDTO;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.sql.SQLException;
@@ -16,6 +17,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 /**
  *
@@ -43,15 +45,26 @@ public class DeletePaymentServlet extends HttpServlet {
         int paymentID = Integer.parseInt(payid);
         String url = "";
         String button = request.getParameter("btAction");
+        HttpSession session = request.getSession(false);
         try {
-            if (button.equals("delete")) {
-                PaymentMethodDAO dao = new PaymentMethodDAO();
+            if (session != null) {
+                UserDetailsDTO dto1 = (UserDetailsDTO) session.getAttribute("USER");
 
-                boolean result = dao.deletePayment(paymentID);
+                if (dto1 != null) {
+                    if (dto1.getRoleID() == 1 || dto1.getRoleID() == 2) {
+                        if (button.equals("delete")) {
+                            PaymentMethodDAO dao = new PaymentMethodDAO();
 
-                if (result) {
-                    url = DELETE_SUCCESS;
+                            boolean result = dao.deletePayment(paymentID);
+
+                            if (result) {
+                                url = DELETE_SUCCESS;
+                            }
+                        }
+                    }
                 }
+           }else{
+            url = "erorr.jsp";
             }
         } catch (SQLException ex) {
             log("UpdatePaymentServlet SQL: " + ex.getMessage());
